+++ frontend/src/app/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { userProfileApi } from '../entities/header-profile/userProfileApi';
+import { userProfileApi } from '../entities/user/Profile/userProfileApi';
 import { userLogin } from '../entities/user/Auth/loginApi';
 import { getTweetApi } from '../features/twit/twitApi';
