+++ frontend/src/entities/header-profile/header-profile.tsx
@@ -1,6 +1,6 @@
 import IUserProfile from '../../shared/types/IUserProfile';
 import HeaderDesription from './header-description';
-import { useGetUserProfileQuery } from './userProfileApi';
+import { useGetUserProfileQuery } from '../user/Profile/userProfileApi';
 import StickyHeader from './sticky-header';
 import Spiner from '../../shared/IU/spiner/spiner';
 import { useGetTweetsQuery } from '../../features/twit/twitApi';
@@ -24,6 +24,7 @@ export default function HeaderProfile(
     following,
     avatar,
   }: IUserProfile = data;
+
   return (
     <div className="flex flex-row flex-wrap border-b">
       <StickyHeader name={name} />
